+++ R/zzz.R
@@ -11,8 +11,8 @@
   ## Print on Screen
   packageStartupMessage("**-----------------------------------------------------------------**")
   packageStartupMessage("** maotai")
-  packageStartupMessage("**")
   packageStartupMessage("**  - Tools for Matrix Algebra, Optimization and Inference Problems")
+  packageStartupMessage("**")
   packageStartupMessage("** Version    : ",this.version,"      (",this.year,")",sep="")
   packageStartupMessage("** Maintainer : Kisung You (kyoustat@gmail.com)")
   packageStartupMessage("**")